rename span file
