+++ lib/lightstep/tracer/span.rb
